+++ test/map_forms_splice_SUITE.erl
@@ -203,5 +203,6 @@ spec_form(Pos, Name) ->
 run_splice(Handler, Forms, InitState) ->
     astranaut_return:run(
       astranaut_traverse:run(
-        astranaut:map_forms_splice(Handler, Forms, #{traverse => none}),
+        astranaut_macro:map_forms_splice(
+          Handler, Forms, #{traverse => none}),
         ?MODULE, #{}, InitState)).